requirements.txt
Django==5.2.1
psycopg2-binary==2.9.10
requests==2.31.0
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0
setuptools>=65.5.1
wheel>=0.38.4
pip>=25.1.1 